requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
iniconfig==2.3.0
mysql-connector-python==9.5.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
